# [perf] fix perf regression from #12253